ovms-api/app/requirements.txt
pytest-cov==4.0.0 ; python_version >= "3.9" and python_version < "4.0"
pytest==7.4.3 ; python_version >= "3.9" and python_version < "4.0"
pandas==2.1.3
uvicorn
fastapi
python-multipart
pillow
numpy
opencv-python
tritonclient[http]